pdf-parsing/requirements.txt
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
sentence-transformers==2.7.0
google-generativeai==0.3.2
requests==2.31.0
torch==2.0.1
torchvision==0.15.2
transformers==4.34.0
huggingface-hub==0.16.4
tokenizers==0.14.0
